Extract and log the action for the commit
Riak only has a concept of `store` or `delete`. When deleting the
post-commit hook receives an object with an empty value and a metadata
header of `X-Riak-Deleted`.

With this update we detect `X-Riak-Deleted` and use it to set an
`Action` variable with either `store` or `delete`. Also the JSON parsing
needed a small update to gracefully handle the result when asked to
parse an empty value.

Fixes #1.