Don't JSON decode the riak object value
There's no need to do that! We can just pass the object value as-is into
the kafka message. The Kafka message is also JSON encoded but that just
means any consumer interested in parsing the riak object value will need
to parse the kafka message and then parse the value key.

Leaving that extract parsing to the consumer makes the post-commit hook
much simpler and more reliable since JSON decoding has a wide range of
potential error states.

It's also entirely possible that we'll have some riak buckets that don't
have JSON encoded values. Just passing along the Riak object value as
we're given seems the most truthy for a commit log.

Fixes #2