Reduce scope of the hook: only forward to OTP
Instead of doing all the work in a single Erlang function we're adding
an OTP application that will handle the bulk of the effort to reliably
produce to Kafka.

Rather than integrate an OTP application into the running Riak node and
Erlang version (which has proven difficult and complex) this hook will
now only send the action, bucket, key, and value from the Riak object to
a separate OTP application.

The only tricky part of this hook is then to know where to send the Riak
object data. Currently it's doing a naive transformation of its own node
name and expecting the external OTP service to be separately running.