Send client timing to statsd
We're making the connect to statsd directly with `gen_udp` so we can
avoid loading up with hook with a dependency so the deployment is just
the one BEAM file. Each run of the hook will make a new UDP socket
connection to statsd but we think that'll be fine since it's only
connecting locally.

Fixes #1