Rename the module to postcommit_hook
Naming the module kafka_riak_commitlog was a legacy of when the hook was
going to be the entire process. Thankfully those crazy days are behind
us and the hook will only be forwarding the information on to a robust
OTP application that will be responsible for writing to Kafka.

Fixes #5