Update for new dev-services setup
With dev-services we're running Riak (and other services) in Docker
containers instead of locally on the mac. That means we need to change
how this hook installs itself for development.

There are two major things that are scripted for development:

1. Compiling the beam file and placing it into Riak
2. Configuring Riak buckets to call the postcommit hook

For 1: Instead of controlling Riak directly we'll now compile the
postcommit hook beam file and copy it to the right directory in
dev-services.  Dev-services will know how to place the hook into Riak if
the file is present.

For 2: The script now knows how to talk to the new dev-services instance
of Riak to configure the buckets.

Fixes #15