Get Commitlog hostname from environment var
Currently, Post-commit Hook (PCH) and Commitlog (CL) run within the CDE
on every Riak node and communicate with each other directly via Erlang
messaging without hitting the network. A consequence of this design is
that updates to CL require that we take down every Riak node (which is
a time-consuming manual process). As part of the AWS migration we want
to decouple CL from Riak / PCH, so they can be upgraded and scaled
independently. This requires configuring PCH so it can find CL on the
network.

Going forward, PCH will retrieve the CL hostname from an environment
variable, `COMMITLOG_HOSTNAME`, falling back to the Erlang host returned
by `node().` if the env var is not set. In other words, unless the env
var is set, PCH will expect CL to run on the same host as Riak; e.g. for
dev-services.

- defaults to the Erlang host from `node().`
- log bucket and key for successful calls to Commitlog
- update `make help` output
- add .tool-versions