Retry failed calls to Commitlog
With the migration to AWS we are separating Commitlog from Riak and PCH.
Commitlog will run on a separate machine. As a result, we are
introducing a network connection between PCH and Commitlog, so we need
the ability to handle short and long term Commitlog disruptions.

In the event of a short term Commitlog disruption (e.g. during
a deployment of a new version of Commitlog), PCH will retry failed calls
to Commitlog using exponential backoff. If all attempts fail (the
maximum is configurable), a message is logged at the WARN level that can
be used for recovery / replay.

In the event of a long term Commitlog disruption (e.g. due to
a downstream issue with Heroku Kafka), a Circuit Breaker process is
started that disables retries until a configurable amount of time
elapses. This allows PCH to fail fast rather than apply excessive back
pressure on Riak (due to retries with exponential backoff).

The vast majority of the time, when Commitlog is fully available, the
Circuit Breaker does not run and retries are not needed.

Refactoring:
- move defines back into postcommit_hook.erl
- add `shell` target to Makefile
- rename call_commitlog to send_to_commitlog
- rename do_call_commitlog to call_commitlog
- send_timing_to_statsd returns error
- test query_circuit_breaker_state error cases