Retry failed calls to Commitlog
With the migration to AWS we are separating Commitlog from Riak and PCH.
Commitlog will run on a separate machine. As a result, we are
introducing a network connection between PCH and Commitlog, so we need
the ability to handle transient Commitlog disruptions.

This commit adds retry logic to PCH and uses an exponential backoff
algorithm to retry failed calls to Commitlog up to a configurable number
of attempts.

Refactoring:
- rename RetryDelay to Delay
- minor refactor
- add jitter_test_
- add retry_send_to_commitlog_test_
- increment Attempts in send_to_commitlog_with_retries
- minor refactor
- tests do not include postcommit_hook.hrl
- minor changes
- unwrap gen_server errors
- fix jitter test